+++ backend/src/index.ts
@@ -1,8 +1,18 @@
-// import { OpenAPIHandler } from "@orpc/openapi/fetch";
-// import { OpenAPIReferencePlugin } from "@orpc/openapi/plugins";
+/**
+ * Application entry point.
+ *
+ * Wires together the Hono app, middleware, and route handlers:
+ *
+ * - CORS is applied once via a shared config factory and mounted on `/api/*`.
+ * - The better-auth session is injected into Hono context for all v1 routes
+ *   so downstream middleware and oRPC procedures can read it without hitting
+ *   the auth layer again.
+ * - Stripe webhook routes bypass oRPC because Stripe signature verification
+ *   requires the raw, unconsumed request body.
+ * - All other API calls are handled by the oRPC `RPCHandler`.
+ */
 import { onError } from "@orpc/server";
 import { RPCHandler } from "@orpc/server/fetch";
-// import { Scalar } from "@scalar/hono-api-reference";
 import { Hono } from "hono";
 import { cors } from "hono/cors";
 import { logger } from "hono/logger";
@@ -19,36 +29,24 @@ type SessionVariables = {
   } | null;
 };
 
-const app = new Hono<{ Variables: SessionVariables }>();
-
-app.use(
-  "*",
-  logger()
-  // logger({
-  //   pino: logger,
-  //   http: {
-  //     reqId: () => Bun.randomUUIDv7(),
-  //   },
-  // })
-);
-
-app.use(
-  "/api/*",
-  cors({
+/**
+ * Builds the shared CORS middleware options.
+ *
+ * Allows requests from the configured production and local origins.
+ * Falls back to `LOCAL_CORS_ORIGIN` for any unlisted origin so the app
+ * remains usable in development without broad wildcard permissiveness.
+ */
+function buildCorsOptions() {
+  return cors({
     origin: (origin) => {
-      // If the request comes from any of these safe origins, allow it!
       const allowedOrigins = [
         env.PROD_CORS_ORIGIN,
         env.PROD_SUB_CORS_ORIGIN,
         env.LOCAL_CORS_ORIGIN,
       ];
-
-      // Return the exact origin the browser asked for if it's in our safe list
       if (origin && allowedOrigins.includes(origin)) {
         return origin;
       }
-
-      // Fallback
       return env.LOCAL_CORS_ORIGIN;
     },
     allowHeaders: [
@@ -61,44 +59,22 @@ app.use(
     exposeHeaders: ["Content-Length", "X-Kuma-Revision"],
     maxAge: 600,
     credentials: true,
-  })
-);
+  });
+}
+
+const app = new Hono<{ Variables: SessionVariables }>();
+
+app.use("*", logger());
+
+// Apply CORS to all /api/* routes using the shared config
+app.use("/api/*", buildCorsOptions());
 
 app.get("/", (c) => c.json({ message: "Hello Hono!", status: "ok" }));
 
 const v1 = new Hono<{ Variables: SessionVariables }>();
 
-v1.use(
-  "*",
-  cors({
-    origin: (origin) => {
-      // If the request comes from any of these safe origins, allow it!
-      const allowedOrigins = [
-        `${env.PROD_CORS_ORIGIN}`,
-        `${env.PROD_SUB_CORS_ORIGIN}`,
-        `${env.LOCAL_CORS_ORIGIN}`,
-      ];
-
-      // Return the exact origin the browser asked for if it's in our safe list
-      if (origin && allowedOrigins.includes(origin)) {
-        return origin;
-      }
-
-      // Fallback
-      return env.LOCAL_CORS_ORIGIN;
-    },
-    allowHeaders: [
-      "Content-Type",
-      "Authorization",
-      "X-Custom-Header",
-      "Upgrade-Insecure-Requests",
-    ],
-    allowMethods: ["POST", "GET", "OPTIONS"],
-    exposeHeaders: ["Content-Length", "X-Kuma-Revision"],
-    maxAge: 600,
-    credentials: true,
-  })
-);
+// Apply CORS to all v1 routes as well (v1 is mounted under /api/v1)
+v1.use("*", buildCorsOptions());
 
 // Stripe webhook — must stay as a raw Hono route: Stripe signature verification
 // requires the unparsed request body, which oRPC's handler would consume first.
@@ -112,29 +88,30 @@ v1.get("health", (c) =>
   })
 );
 
-// Inject better-auth session into Hono context for all v1 routes
-// Unsure why got CORS errors
+// Inject the better-auth session into Hono context for all v1 routes so
+// oRPC procedures can access it via `context.session` without re-fetching.
 v1.use("*", async (c, next) => {
   const session = await auth.api.getSession({ headers: c.req.raw.headers });
-
   if (session?.user) {
     c.set("session", { userId: session.user.id });
   }
-
   await next();
 });
 
 // better-auth sign-in / sign-up / session routes
 v1.on(["POST", "GET"], "/auth/**", (c) => auth.handler(c.req.raw));
 
-// oRPC RPCHandler — serves all procedures via RPC protocol (matches RPCLink on frontend).
-// Previously used OpenAPIHandler but switched due to protocol mismatch:
-// RPCLink sends { json, meta } envelope which OpenAPIHandler doesn't understand.
+/**
+ * oRPC RPCHandler — serves all procedures via the RPC protocol.
+ *
+ * RPCLink on the frontend sends `{ json, meta }` envelopes, which this
+ * handler understands natively. Validation errors are logged in detail to
+ * aid debugging without leaking them to the client.
+ */
 const rpcHandler = new RPCHandler(appRouter, {
   interceptors: [
     onError((error) => {
       pinoLogger.error(error, "oRPC unhandled error");
-      // Log validation errors in detail for debugging
       if (
         error instanceof Error &&
         error.cause &&
@@ -152,42 +129,7 @@ const rpcHandler = new RPCHandler(appRouter, {
   ],
 });
 
-// const openAPIHandler = new OpenAPIHandler(appRouter, {
-//   plugins: [
-//     new OpenAPIReferencePlugin({
-//       specPath: "/openapi.json",
-//       specGenerateOptions: {
-//         info: {
-//           title: "OneCurrency API",
-//           version: "1.0.0",
-//           description: "API documentation for OneCurrency e-wallet application",
-//         },
-//         servers: [
-//           {
-//             url: `http://localhost:${env.API_PORT}/api/v1`,
-//             description: "Local development server",
-//           },
-//         ],
-//       },
-//     }),
-//   ],
-//   interceptors: [
-//     onError((error) => {
-//       pinoLogger.error(error, "oRPC unhandled error");
-//       if (error.cause && typeof error.cause === "object") {
-//         const cause = error.cause;
-//         if ("flatten" in cause || "issues" in cause) {
-//           pinoLogger.error(
-//             { validationError: cause },
-//             "Input validation failed - detailed error"
-//           );
-//         }
-//       }
-//     }),
-//   ],
-// });
-
-// RPCHandler route — no body proxy needed, RPCHandler reads the raw request directly
+// Mount the oRPC handler on all unmatched v1 routes
 v1.use("*", async (c, next) => {
   const context: ORPCContext = { session: c.get("session") ?? null };
 
@@ -203,49 +145,7 @@ v1.use("*", async (c, next) => {
   await next();
 });
 
-// Body proxy + OpenAPIHandler route (commented out — kept for reference/rollback)
-// const BODY_PARSER_METHODS = new Set([
-//   "arrayBuffer",
-//   "blob",
-//   "formData",
-//   "json",
-//   "text",
-// ] as const);
-//
-// type BodyParserMethod = typeof BODY_PARSER_METHODS extends Set<infer T>
-//   ? T
-//   : never;
-//
-// v1.use("*", async (c, next) => {
-//   const request = new Proxy(c.req.raw, {
-//     get(target, prop) {
-//       if (BODY_PARSER_METHODS.has(prop as BodyParserMethod)) {
-//         return () => c.req[prop as BodyParserMethod]();
-//       }
-//       const value = Reflect.get(target, prop, target);
-//       // Bind functions to preserve this-context (e.g., request.clone())
-//       return typeof value === "function" ? value.bind(target) : value;
-//     },
-//   });
-//
-//   const context: ORPCContext = { session: c.get("session") ?? null };
-//
-//   const { matched, response } = await openAPIHandler.handle(request, {
-//     prefix: "/api/v1",
-//     context,
-//   });
-//
-//   if (matched) {
-//     return c.newResponse(response.body, response);
-//   }
-//
-//   await next();
-// });
-
 app.route("/api/v1", v1);
-
-// Scalar API Reference UI (served by oRPC OpenAPIReferencePlugin at /api/v1/openapi.json)
-// app.get("/api/v1/docs", Scalar({ url: "/api/v1/openapi.json" }));
 
 export default {
   port: env.API_PORT,
